+++ bitacross-worker/service/src/worker_peers_updater.rs
